refactor(engine): extract AROfficeProcessor (Phase 2 PR-5)
